Data loading (#1)
* [WIP] Image captioning with VGG16 and LSTM

* Broken af image loader using TFRecords

* Fix data loader misclassifying input images

When I was scraping the dataset directory initially, I was looking for the word
"cat" in each of the input image filenames. However, the input images were in a
directory with the word "cat" in it, so it was labelling every image as a cat.
It works now when the dataset directory has a captial "C" - not really a fix.